+++ src/components/dropdown.stories.tsx
@@ -1,8 +1,12 @@
-import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Button } from "./button";
 import { Dropdown } from "./dropdown";
+import { DropdownContent } from "./dropdown_content";
+import { DropdownDivider } from "./dropdown_divider";
+import { DropdownItem } from "./dropdown_item";
+import { DropdownMenu } from "./dropdown_menu";
+import { DropdownTrigger } from "./dropdown_trigger";
 
 const meta = {
   title: "Dropdown",
@@ -16,19 +20,19 @@ type Story = StoryObj<typeof meta>;
 export const BasicStory: Story = {
   render: (args) => (
     <Dropdown {...args}>
-      <Dropdown.Trigger>
+      <DropdownTrigger>
         <Button>Dropdown</Button>
-      </Dropdown.Trigger>
-      <Dropdown.Menu>
-        <Dropdown.Content>
-          <Dropdown.Item>Lorem Ipsum</Dropdown.Item>
-          <Dropdown.Item>Lorem Ipsum</Dropdown.Item>
-          <Dropdown.Divider />
-          <Dropdown.Item active>Lorem Ipsum</Dropdown.Item>
-          <Dropdown.Divider />
-          <Dropdown.Item>Lorem Ipsum</Dropdown.Item>
-        </Dropdown.Content>
-      </Dropdown.Menu>
+      </DropdownTrigger>
+      <DropdownMenu>
+        <DropdownContent>
+          <DropdownItem>Lorem Ipsum</DropdownItem>
+          <DropdownItem>Lorem Ipsum</DropdownItem>
+          <DropdownDivider />
+          <DropdownItem active>Lorem Ipsum</DropdownItem>
+          <DropdownDivider />
+          <DropdownItem>Lorem Ipsum</DropdownItem>
+        </DropdownContent>
+      </DropdownMenu>
     </Dropdown>
   ),
 };