+++ src/components/file_input.test.tsx
@@ -1,6 +1,6 @@
 import { renders } from "../support";
 
-import { FileInput } from "./form_file_input";
+import { FileInput } from "./file_input";
 
 describe(FileInput, () => {
   renders(FileInput, ".file-input");