+++ src/components/form.test.tsx
@@ -1,7 +1,0 @@
-import { renders } from "../support";
-
-import { Form } from "./form";
-
-describe(Form, () => {
-  renders(Form, "form");
-});