+++ src/components/hero.stories.tsx
@@ -1,8 +1,10 @@
-import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Container } from "./container";
 import { Hero } from "./hero";
+import { HeroBody } from "./hero_body";
+import { HeroFoot } from "./hero_foot";
+import { HeroHead } from "./hero_head";
 import { Subtitle } from "./subtitle";
 import { Title } from "./title";
 
@@ -18,14 +20,14 @@ type Story = StoryObj<typeof meta>;
 export const BasicStory: Story = {
   render: (args) => (
     <Hero {...args}>
-      <Hero.Head>Lorem Ipsum</Hero.Head>
-      <Hero.Body>
+      <HeroHead>Lorem Ipsum</HeroHead>
+      <HeroBody>
         <Container>
           <Title>Lorem Ipsum</Title>
           <Subtitle>Lorem Ipsum</Subtitle>
         </Container>
-      </Hero.Body>
-      <Hero.Foot>Lorem Ipsum</Hero.Foot>
+      </HeroBody>
+      <HeroFoot>Lorem Ipsum</HeroFoot>
     </Hero>
   ),
 };