+++ src/components/media.stories.tsx
@@ -1,10 +1,12 @@
-import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Content } from "./content";
 import { Delete } from "./delete";
 import { Image } from "./image";
 import { Media } from "./media";
+import { MediaContent } from "./media_content";
+import { MediaLeft } from "./media_left";
+import { MediaRight } from "./media_right";
 
 const meta = {
   title: "Media",
@@ -20,18 +22,18 @@ const SAMPLE_URL = "https://source.unsplash.com/BDIcmo7netA/96x96";
 export const BasicStory: Story = {
   render: (args) => (
     <Media {...args}>
-      <Media.Left>
+      <MediaLeft>
         <Image square dimensions={48} src={SAMPLE_URL} />
-      </Media.Left>
-      <Media.Content>
+      </MediaLeft>
+      <MediaContent>
         <Content>
           <strong>Ringo</strong> <span>@ringo</span>
           <p>The quick brown fox jumped over the lazy dog.</p>
         </Content>
-      </Media.Content>
-      <Media.Right>
+      </MediaContent>
+      <MediaRight>
         <Delete />
-      </Media.Right>
+      </MediaRight>
     </Media>
   ),
 };