+++ src/components/panel.stories.tsx
@@ -1,10 +1,13 @@
-import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faStar } from "@fortawesome/free-solid-svg-icons";
 
 import { Panel } from "./panel";
+import { PanelHeading } from "./panel_heading";
+import { PanelTabs } from "./panel_tabs";
+import { PanelBlock } from "./panel_block";
+import { PanelIcon } from "./panel_icon";
 
 const meta = {
   title: "Panel",
@@ -18,17 +21,17 @@ type Story = StoryObj<typeof meta>;
 export const BasicStory: Story = {
   render: (args) => (
     <Panel {...args}>
-      <Panel.Heading>Lorem Ipsum</Panel.Heading>
-      <Panel.Tabs>
+      <PanelHeading>Lorem Ipsum</PanelHeading>
+      <PanelTabs>
         <a>Like</a>
         <a>Save</a>
-      </Panel.Tabs>
-      <Panel.Block>
-        <Panel.Icon>
+      </PanelTabs>
+      <PanelBlock>
+        <PanelIcon>
           <FontAwesomeIcon icon={faStar} />
-        </Panel.Icon>
+        </PanelIcon>
         Lorem Ipsum
-      </Panel.Block>
+      </PanelBlock>
     </Panel>
   ),
 };